+++ NEWS.md
@@ -1,4 +1,11 @@
-# GCAManalysis 0.1.0
+# GCAM Analysis Package
+
+## version 0.1.1
+
+* Give the structure returned by the trend analysis a class and some
+  methods. 
+
+## version 0.1.0
 
 * Add trend analysis function.